Move to v0.1.1 and add __version__
The version number is now 0.1.1. It is available via the __version__
special attribute and kept in sync with setup.py.

Because we do not want to inport panedr in setup.py (due to uninstalled
requirements that would prevent the import), we use the panedr/VERSION
file to keep the version in sync between setup.py and __version__. The
version number is only written once, in panedr/VERSION.