Move to a more object oriented architecture
This commit introduces the `EDRFile` class that wraps all the steps to
read an EDR file. `do_enxnms`, `do_eheader`, and `go-enx` are now method
of that class. An EDRFile object can also be iterated on to read the
frames.

This commit also fixes a bug on the last line of the progress meter. The
frame number displayed as last read was one frame to much.